Allow omission of network card name for SCALE
Starting in Electric Eel we reduce the amount of spam being
written to the console during boot. This means that the console
parser will be unable to determine network interface name.

This commit slightly reworks the library so that we can clone
it into jenkins dynamically and run automation from the latest
code.